+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.0.0-beta.4
+
+## New Features
+
+- Adding ability to specify your own instance of an `apiCheck` which allows you to use your own checkers.
+
+## Bug Fixes
+
+- Fixing the prefix for the `scopeTypes` instance of an `apiCheck` to be `angular-scope-types` rather than `api-check-angular`
+
 # 1.0.0-beta.3
 
 ## Bug Fixes